Interface-hosted inh attributes: disabled controversial tests
In cases where `getChild().attr()` broadcasting happens via
multiple separate paths in the inheritance tree, the
priority rules are murky.  For now, the semantics for the
following case are UNDEFINED in the sense that client code
must not rely on the current behaviour and that these
cases may be disallowed in the future or that their semantics
may change:

------------------------------
  // A, B: arbitrary AST node types
  // such that B <: A (may be indirect)

  // Example:
  A ::= ...
  B : A ::= ...
------------------------------
  inh T ASTNode.attr();
  // ASTNode, T, attr() may be any suitable type, node type,
  // and attribute name; attr may be parametric, circular etc.

  interface I { ... }
  eq A.getChild().attr() = ...
  eq I.getChild().attr() = ...
  // analogously for getChild(int)

  B implements I;  // may be indirect
  // where A </: I

  // In other words, B inherits equations for
  //   getChild().attr()
  // from two distinct paths

  // The following leads to UNDEFINED behaviour:
  eq S.getX().attr() = ...
  // where S <> B and either:
  // 1.  B <: S <: A
  //  or
  // 2.  B <: S <: I
------------------------------

Note that the corresponding case for two paths of
interface inheritance is already disallowed.