Update test scripts
Made the gradlew script in the test directory have more or less feature
parity with the ant script. It now supports `-Dtest` and `-DtestSuite`.

Removed the hard-coded "jastadd2.jar" references in the test suite,
replaced by a single property from gradle/ant.
This property has been prefixed with `../`, meaning it is no longer
necessary to copy jastadd2.jar from the root down into the test dir.

Updated the root README to reflect that jastadd-test is no longer a
submodule.